https://coveord.atlassian.net/browse/CDX-912

+++ packages/cli-e2e/utils/login.ts
@@ -1,11 +1,6 @@
 import retry from 'async-retry';
 import type {Browser, Page, Target} from 'puppeteer';
-import {
-  answerPrompt,
-  CLI_EXEC_PATH,
-  getConfigFilePath,
-  isGenericYesNoPrompt,
-} from './cli';
+import {answerPrompt, getConfigFilePath, isGenericYesNoPrompt} from './cli';
 import LoginSelectors from './loginSelectors';
 import {strictEqual} from 'assert';
 import {readJSON, writeJSON, existsSync} from 'fs-extra';
@@ -66,16 +61,13 @@ async function possiblyAcceptCustomerAgreement(page: Page) {
 
 export function runLoginCommand(orgId: string, env: string) {
   const args: string[] = [
-    CLI_EXEC_PATH,
+    process.env.CLI_EXEC_PATH!,
     'auth:login',
     `-e=${env}`,
     `-o=${orgId}`,
   ];
-  if (process.platform === 'win32') {
-    args.unshift('node');
-  }
   const loginTerminal = new Terminal(
-    args.shift()!,
+    'node',
     args,
     undefined,
     global.processManager!,
@@ -131,7 +123,7 @@ async function startLoginFlow(browser: Browser) {
   });
   await Promise.all([
     page.click(`${LoginSelectors.passwordView} ${LoginSelectors.SubmitInput}`),
-    page.waitForNavigation({waitUntil: 'networkidle2'}),
+    page.waitForNavigation({waitUntil: 'networkidle2', timeout: 2 * 60e3}),
   ]);
 
   await staySignedIn(page);
@@ -171,5 +163,22 @@ export async function clearAccessTokenFromConfig() {
   }
   const cfg = await readJSON(getConfigFilePath());
   delete cfg.accessToken;
+  await writeJSON(getConfigFilePath(), cfg);
+}
+
+export async function loginWithApiKey(
+  apiKey: string,
+  orgId: string,
+  env: string
+) {
+  if (!existsSync(getConfigFilePath())) {
+    throw 'Missing config file';
+  }
+  const cfg = await readJSON(getConfigFilePath());
+  cfg.accessToken = apiKey;
+  cfg.organization = orgId;
+  cfg.environment = env;
+  cfg.analyticsEnabled = false;
+  cfg.anonymous = true;
   await writeJSON(getConfigFilePath(), cfg);
 }
